Allow `git describe` to function without tags
Closes #12

Note that --always must be paired with --tags to get the expected output
in the most common use case when there are tags. Without it only the
short hash is ever returned. Combined with --tags the output is the same
as previous when tags are present but will fall back to identifying the
commit by its hash if no tags are found.